build(core)!: switch library type to INTERFACE
Transitioned `digitalcurling3::core` from a STATIC library to an INTERFACE library.
This change converted `src/version.cpp.in` to `include/digitalcurling3/detail/version_info.hpp.in`, and removed the `src/` directory as it is no longer required.

BREAKING CHANGE: The build type of core library has changed from STATIC to INTERFACE. It is now a header-only library.